requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.12.1
numpy==2.3.2
pypdf==5.9.0
python-docx==1.2.0
sentence-transformers==5.2.3
faiss-cpu==1.13.2
google-genai==1.56.0
python-dotenv==1.1.1
langchain==1.2.12
langgraph==1.1.1
langchain-google-genai==4.2.1
